start of ts7200.rs cleanup + reorg
in this episode, we split the gdbstub impl into it's own file, and
shuffle around the directory structure a bit
